requirements.txt
boto3==1.14.51
botocore==1.17.51
dj-database-url==0.5.0
Django==1.11.29
django-ckeditor==5.9.0
django-forms-bootstrap==3.1.0
django-js-asset==1.2.2
django-storages==1.9.1
django-tinymce==3.0.2
docutils==0.15.2
gunicorn==20.0.4
heroku==0.1.4
jmespath==0.10.0
npm==0.1.1
optional-django==0.1.0
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
stripe==2.50.0
